+++ packages/guide/src/app/shared/components/database-item-details/buy-at/buy-at.component.ts
@@ -1,13 +1,12 @@
 import { Component, input } from '@angular/core';
-import { Consumable, DatabaseItem } from "@ci/data-types";
-import { AddSpacesToPascalCasePipe } from "../../../pipes/add-spaces-to-pascal-case.pipe";
+import { DatabaseItem } from "@ci/data-types";
 import { TranslatePipe } from "@ngx-translate/core";
 import { RouterLink } from "@angular/router";
 import { MoneyComponent } from "../../money/money.component";
 
 @Component({
     selector: 'app-buy-at',
-    imports: [AddSpacesToPascalCasePipe, TranslatePipe, RouterLink, MoneyComponent],
+    imports: [TranslatePipe, RouterLink, MoneyComponent],
     templateUrl: './buy-at.component.html'
 })
 export class BuyAtComponent {